Open start page automatically when starting lobby
Use server-sent events to automatically open the start page when the
host starts the lobby. Since they are one-sided, keep them alive as long
the lobby is alive. If an extension joins a different lobby, it
automatically closes its current eventstream (which the server wont know
about) and wont receive any updates from the old lobby.

Signed-off-by: Victor Löfgren <victor.lofgren99@pm.me>