Reference latest docker image instead of main
Main will be the last push to the main branch, while latest is the
latest tagged release. You usually want to update the server in
conjuction with browser extension, otherwise they might be incompatible.
Its a lot safer to reference the latest release, although there is a
short window where one of them (server/extension) might be updated
before the other during a release. Dont know what to do about that.

Signed-off-by: Victor Löfgren <victor.lofgren99@pm.me>