Implement simple library to handle settings
Chrome only allows one service worker, instead of multiple background
scripts like Firefox. So to remedy that I made the settings file and
background scripts into module. This makes it possible for the
background script to import it. It also needed to be imported in a
special way in the other files as well...

Signed-off-by: Victor Löfgren <victor.lofgren99@pm.me>